+++ kernel/detect.c
@@ -17,16 +17,15 @@
 * License along with NOS.  If not, see <http://www.gnu.org/licenses/>.  *
 ************************************************************************/
 
-/*@file version.h
+/*@file detect.c
 * @author Ahmad Dajani <eng.adajani@gmail.com>
-* @date 5 Jan 2020
-* @brief Kernel version header file
+* @date 30 Jan 2020
+* @brief Basic hardware detection
 */
 
-#ifndef __VERSION_H
-    #define __VERSION_H
-    #define MAJOR_VERSION 0
-    #define MINOR_VERSION 8
-    #define CODE_NAME "Hypsilophodon"
-    #define COPY_RIGHT "(c)2020 By Ahmad Dajani, eng.adajani@gmail.com"
-#endif
+#include <conio.h> /* printFormat */
+#include <kernel/detect.h> /*  */
+
+void detectHardware(void) {
+    //TODO
+}